avoid unnecessary compare
We walk on the higher level until we found a key greater than target key, then go to the lower level.
There is 1/3 chance we will compare the same node of the upper level.

After this optimization, when the skiplist size is 1MB, the improvement is about 5%, but when the size is larger like 64MB, the improvement is only 2%
May be that's because the key is already in L1 cache when we compare it again.